rename sphere
